refactor(payload): split admin/adapters.ts into adapters/ directory
Separates router contracts from cookie contracts:
- adapters/router.ts: RouterAdapterComponent, RouterAdapterRouter, LinkAdapterProps
- adapters/cookies.ts: CookieStore, CookieOptions
- adapters/index.ts: barrel re-export

Makes room for ServerAdapter to land as adapters/server.ts on its own
branch without re-mingling concerns in a single file.
