+++ packages/payload/src/admin/adapters/index.ts
@@ -1,0 +1,2 @@
+export type * from './cookies.js'
+export type * from './router.js'
